src/requirements.txt
cryptography==46.0.5
email-validator==2.3.0
fastapi==0.133.0
firebase_admin==7.1.0
gunicorn==25.1.0
httpx==0.28.1
jinja2==3.1.6
locust==2.43.3
pytest==9.0.2
python-dotenv==1.2.1
python-multipart==0.0.22
psycopg==3.3.3
psycopg-binary==3.3.3
SQLAlchemy==2.0.47
uvicorn==0.41.0

